src/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "vlm vllm pipeline backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "modal>=1.3.1",
    "numpy>=2.2.6",
    "openai>=2.15.0",
    "pandas>=2.3.3",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.9",
    "python-dotenv>=1.2.1",
    "vllm>=0.13.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
